requirements.txt
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.1
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
blinker==1.8.2
Brlapi==0.8.3
Brotli==1.0.9
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distro==1.7.0
Django==4.2.11
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
executing==2.0.1
eyeD3==0.8.10
filelock==3.6.0
Flask==3.0.3
Flask-WTF==1.2.1
fpdf==1.7.2
httplib2==0.20.2
idna==3.7
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
more-itertools==8.10.0
mutagen==1.45.1
nemo-emblems==6.0.1
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
onboard==1.4.1
packaging==21.3
PAM==0.4.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.43
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
Pygments==2.18.0
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu3
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-gnupg==0.4.8
python-magic==0.4.24
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
reportlab==3.6.8
requests==2.31.0
requests-file==1.5.1
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.5.0
stack-data==0.6.3
systemd-python==234
texttable==1.6.4
tinycss2==1.1.1
tldextract==3.1.2
traitlets==5.14.3
typing_extensions==4.11.0
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==2.2.1
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.2.3
websockets==9.1
Werkzeug==3.0.3
WTForms==3.1.2
xdg==5
xkit==0.0.0
xlrd==1.2.0
yt-dlp==2022.4.8
zipp==1.0.0
